refactor: remove first and last name from registration
