+++ packages/common/src/types/authentication.ts
@@ -1,8 +1,6 @@
 import { z } from "zod";
 
 export const ZAuthRegistration = z.object({
-  firstName: z.string(),
-  lastName: z.string(),
   email: z.string().email(),
   srp: z.object({
     salt: z.string(),
